Automated_Data_Analysis_0.1
I've yet to implement the analysis of more specific parts of the flow cytometry data, aiming to first make the code work on the more generic data. It went well and works, except that I am running into issues when it comes to adding tables of data to the outputted pdf, although it does work when I make their own separate document. As of now, the code does not add the summary statistics table.

I'm also looking into setting up an extra layer through which we can track the data from each of the various data files and compare them, although I am unsure whether that would be applicable to data sets beyond those from the same lab. Perhaps going through the data of the the files and noting those where the same parameters are found, and then taking an average.

I'd appreciate any insight on either of these matters.

Signed-off-by: pratyush-k-singh <146137799+pratyush-k-singh@users.noreply.github.com>